refactor: close last 2 deferred techdebt items + 2 follow-on bugs
Audit of the 206/208 sweep (a26110a) turned up four loose ends:

1. parish-tauri TD-003: ~940-line `.setup()` closure was deferred. Extract
   8 helpers into new `setup` module: `init_screenshot_mode`,
   `bootstrap_inference_provider`, `init_inference_queue`,
   `init_persistence`, `spawn_event_bus_fanin`, `spawn_world_tick`,
   `spawn_inactivity_tick`, `spawn_debug_tick`, `spawn_autosave_tick`.
   `lib.rs` shrinks 2170 -> 1203 lines; behaviour byte-for-byte identical.
   Mirrors `parish-server::session::spawn_session_ticks` pattern.

2. parish-inference TD-015: Windows `taskkill` zero-coverage. Extract pure
   helpers `taskkill_args` and `pid_string` from `OllamaProcess::stop`;
   add 2 cross-platform tests pinning the `/F /T /PID <pid>` invariant
   without needing a Windows host or a Command-mock abstraction.

3. parish-server TD-011: `tests/ws_integration.rs` referenced
   `validate_ws_upgrade` and `WsValidation` symbols that were never
   actually extracted in a26110a. Extract them now from `ws_handler`,
   wire the handler through them, and confirm the 8 integration tests
   pass.

4. apps/ui + parish-core TODO.md bookkeeping: TD-019/TD-020 in apps/ui
   were mis-classified under "Follow-up: deferred" with bodies marked
   "Fixed". Move them into Done. parish-core Discovery note claimed
   items "recorded as Follow-up" while Follow-up was empty; rewrite.

Net: 208/208 closed, zero Open, zero deferred Follow-up across all 15
TODO.md files. Workspace clippy clean with -D warnings; all 2457 tests
pass.
