+++ parish/crates/parish-tauri/src/lib.rs
@@ -8,6 +8,7 @@ pub mod command_registry;
 pub mod commands;
 pub mod editor_commands;
 pub mod events;
+mod setup;
 
 use parish_core::AUTOSAVE_INTERVAL_SECS;
 
@@ -21,11 +22,9 @@ use tokio::task::JoinHandle;
 use tokio_util::sync::CancellationToken;
 
 use parish_core::config::{FeatureFlags, Provider, ProviderConfig};
-use parish_core::debug_snapshot::{DebugEvent, InferenceDebug};
+use parish_core::debug_snapshot::DebugEvent;
 use parish_core::game_mod::PronunciationEntry;
-use parish_core::inference::{
-    AnyClient, InferenceLog, InferenceQueue, new_inference_log, spawn_inference_worker,
-};
+use parish_core::inference::{AnyClient, InferenceLog, InferenceQueue, new_inference_log};
 use parish_core::npc::manager::NpcManager;
 use parish_core::npc::reactions::ReactionTemplates;
 use parish_core::world::transport::TransportConfig;
@@ -898,58 +897,10 @@ pub fn run() {
         .setup(move |app| {
             let handle = app.handle().clone();
 
-            // ── Screenshot mode ───────────────────────────────────────────────
-            // If --screenshot <dir> was passed, capture the UI at 4 times of day
-            // and exit. No background ticks are started in this mode.
+            // Screenshot mode: --screenshot <dir> captures the UI at four
+            // times of day and exits. No background ticks are started.
             if let Some(dir) = screenshot_dir.clone() {
-                let state_ss = Arc::clone(&state);
-                let handle_ss = handle.clone();
-                tauri::async_runtime::spawn(async move {
-                    // Give the WebView time to fully load the frontend.
-                    // In Xvfb + WebKit2 software rendering the JS bundle takes
-                    // ~15–20 s to parse, JIT, and complete the initial IPC round-trip
-                    // before onMount data is rendered into the DOM.
-                    tokio::time::sleep(Duration::from_secs(20)).await;
-
-                    // Emit the configured theme once so the frontend has a palette
-                    // painted before the first capture.
-                    {
-                        let palette = state_ss.theme_palette.clone();
-                        let _ = handle_ss.emit(events::EVENT_THEME_UPDATE, palette);
-                    }
-                    tokio::time::sleep(Duration::from_secs(3)).await;
-
-                    let times: &[(&str, u32)] =
-                        &[("morning", 7), ("midday", 12), ("dusk", 18), ("night", 22)];
-
-                    if let Err(e) = std::fs::create_dir_all(&dir) {
-                        tracing::warn!(path = %dir.display(), error = %e, "failed to create screenshot dir");
-                    }
-
-                    for (name, target_hour) in times {
-                        // Advance clock to target hour
-                        {
-                            use chrono::Timelike;
-                            let mut world = state_ss.world.lock().await;
-                            let current_hour = world.clock.now().hour() as i64;
-                            let delta = ((*target_hour as i64) - current_hour).rem_euclid(24) * 60;
-                            world.clock.advance(delta);
-                        }
-
-                        // Wait for Svelte to re-render and WebKit to commit the frame
-                        tokio::time::sleep(Duration::from_secs(5)).await;
-
-                        // GDK must be called from the GTK main thread; dispatch and await.
-                        let path = dir.join(format!("gui-{}.png", name));
-                        if let Err(e) = dispatch_screenshot(path).await {
-                            tracing::error!(name = %name, error = %e, "screenshot capture failed");
-                        }
-                    }
-
-                    println!("screenshot: all done, exiting");
-                    handle_ss.exit(0);
-                });
-
+                setup::init_screenshot_mode(handle, Arc::clone(&state), dir);
                 return Ok(());
             }
 
@@ -958,945 +909,27 @@ pub fn run() {
             // so tokio::spawn cannot be called directly here — we must go through
             // tauri::async_runtime::spawn, which uses the Tauri-managed tokio handle.
             let state_setup = Arc::clone(&state);
+            let provider_config_setup = provider_config;
+            let inference_config_setup = inference_config_for_spawn;
             tauri::async_runtime::spawn(async move {
-                // ── Bootstrap inference provider ─────────────────────────────
-                // Runs here (inside the async spawn) rather than synchronously in
-                // run() so the Tauri window is open and can receive setup-status /
-                // setup-progress events while Ollama is being installed/started.
+                if !setup::bootstrap_inference_provider(
+                    &handle,
+                    &state_setup,
+                    &provider_config_setup,
+                    &inference_config_setup,
+                )
+                .await
                 {
-                    let progress = TauriProgress {
-                        app: handle.clone(),
-                        state: Arc::clone(&state_setup),
-                    };
-                    progress.with_setup_status(|status| status.record_status("Starting inference provider setup..."));
-                    tracing::info!("Starting inference provider setup...");
-                    let _ = handle.emit(
-                        events::EVENT_SETUP_STATUS,
-                        events::SetupStatusPayload {
-                            message: "Starting inference provider setup...".to_string(),
-                        },
-                    );
-                    match bootstrap_provider(
-                        &provider_config,
-                        &inference_config_for_spawn,
-                        &progress,
-                    )
-                    .await
-                    {
-                        Ok((client, model_name, ollama_process)) => {
-                            *state_setup.client.lock().await = client;
-                            *state_setup.ollama_process.lock().await = ollama_process;
-                            {
-                                let mut config = state_setup.config.lock().await;
-                                config.model_name = model_name;
-                                config.fill_missing_models_from_presets();
-                            }
-                            record_setup_done(&state_setup, true, String::new());
-                            let _ = handle.emit(
-                                events::EVENT_SETUP_DONE,
-                                events::SetupDonePayload {
-                                    success: true,
-                                    error: String::new(),
-                                },
-                            );
-                        }
-                        Err(e) => {
-                            let error = e.to_string();
-                            record_setup_done(&state_setup, false, error.clone());
-                            tracing::error!("Failed to initialise inference provider: {}", error);
-                            let _ = handle.emit(
-                                events::EVENT_SETUP_DONE,
-                                events::SetupDonePayload {
-                                    success: false,
-                                    error,
-                                },
-                            );
-                            // Do not call process::exit — leave the window open so
-                            // the user can read the error in the setup overlay.
-                            return;
-                        }
-                    }
+                    return;
                 }
 
-                // Initialise inference queue now that the tokio runtime is running
-                {
-                    let provider_name = {
-                        let config = state_setup.config.lock().await;
-                        config.provider_name.clone()
-                    };
-                    let any_client: Option<AnyClient> = if provider_name == "simulator" {
-                        Some(AnyClient::simulator())
-                    } else {
-                        let client_guard = state_setup.client.lock().await;
-                        // `state.client` is already an AnyClient — just clone it.
-                        client_guard.as_ref().cloned()
-                    };
-                    if let Some(ac) = any_client {
-                        let (interactive_tx, interactive_rx) =
-                            tokio::sync::mpsc::channel(16);
-                        let (background_tx, background_rx) =
-                            tokio::sync::mpsc::channel(32);
-                        let (batch_tx, batch_rx) = tokio::sync::mpsc::channel(64);
-                        let worker = spawn_inference_worker(
-                            ac,
-                            interactive_rx,
-                            background_rx,
-                            batch_rx,
-                            state_setup.inference_log.clone(),
-                            state_setup.inference_config.clone(),
-                        );
-                        let queue =
-                            InferenceQueue::new(interactive_tx, background_tx, batch_tx);
-                        let mut iq = state_setup.inference_queue.lock().await;
-                        *iq = Some(queue);
-                        drop(iq);
-                        let mut wh = state_setup.worker_handle.lock().await;
-                        *wh = Some(worker);
-                    }
-                }
-
-                // ── Persistence: auto-load or create save file ──────────────
-                {
-                    use parish_core::persistence::Database;
-                    use parish_core::persistence::SaveFileLock;
-                    use parish_core::persistence::picker::{discover_saves, new_save_path};
-                    use parish_core::persistence::snapshot::GameSnapshot;
-
-                    let saves_dir = state_setup.saves_dir.clone();
-
-                    let world = state_setup.world.lock().await;
-                    let saves = discover_saves(&saves_dir, &world.graph);
-                    drop(world);
-
-                    // Find the most recent unlocked save (iterate in reverse).
-                    let unlocked_save = saves.iter().rev().find(|s| !s.locked);
-
-                    if let Some(save) = unlocked_save {
-                        // Acquire the advisory lock before loading.
-                        let lock = SaveFileLock::try_acquire(&save.path);
-                        if lock.is_some() {
-                            *state_setup.save_lock.lock().await = lock;
-                        }
-
-                        // Load the most recent unlocked save file
-                        match Database::open(&save.path) {
-                            Ok(db) => {
-                                // Find the "main" branch or first branch
-                                let branch = db.find_branch("main").ok().flatten().or_else(|| {
-                                    db.list_branches().ok().and_then(|b| b.into_iter().next())
-                                });
-
-                                if let Some(branch) = branch {
-                                    if let Ok(Some((_snap_id, snapshot))) =
-                                        db.load_latest_snapshot(branch.id)
-                                    {
-                                        let mut world = state_setup.world.lock().await;
-                                        let mut npc_mgr = state_setup.npc_manager.lock().await;
-                                        snapshot.restore(&mut world, &mut npc_mgr);
-                                        npc_mgr.assign_tiers(&world, &[]);
-                                        drop(npc_mgr);
-                                        drop(world);
-
-                                        *state_setup.save_path.lock().await =
-                                            Some(save.path.clone());
-                                        *state_setup.current_branch_id.lock().await =
-                                            Some(branch.id);
-                                        *state_setup.current_branch_name.lock().await =
-                                            Some(branch.name.clone());
-                                        tracing::info!(
-                                            "Restored from {} (branch: {})",
-                                            save.filename,
-                                            branch.name
-                                        );
-                                    } else {
-                                        // Save file exists but no snapshots — save initial state
-                                        let world = state_setup.world.lock().await;
-                                        let npc_mgr = state_setup.npc_manager.lock().await;
-                                        let snap = GameSnapshot::capture(&world, &npc_mgr);
-                                        drop(npc_mgr);
-                                        drop(world);
-                                        let _ = db.save_snapshot(branch.id, &snap);
-
-                                        *state_setup.save_path.lock().await =
-                                            Some(save.path.clone());
-                                        *state_setup.current_branch_id.lock().await =
-                                            Some(branch.id);
-                                        *state_setup.current_branch_name.lock().await =
-                                            Some(branch.name);
-                                    }
-                                }
-                            }
-                            Err(e) => {
-                                tracing::warn!("Failed to open save file {}: {}", save.filename, e);
-                            }
-                        }
-                    } else if saves.is_empty() {
-                        // No saves exist — create a new save file
-                        let path = new_save_path(&saves_dir);
-                        let lock = SaveFileLock::try_acquire(&path);
-                        if lock.is_some() {
-                            *state_setup.save_lock.lock().await = lock;
-                        }
-                        match Database::open(&path) {
-                            Ok(db) => {
-                                if let Ok(Some(branch)) = db.find_branch("main") {
-                                    let world = state_setup.world.lock().await;
-                                    let npc_mgr = state_setup.npc_manager.lock().await;
-                                    let snap = GameSnapshot::capture(&world, &npc_mgr);
-                                    drop(npc_mgr);
-                                    drop(world);
-                                    let _ = db.save_snapshot(branch.id, &snap);
-
-                                    *state_setup.save_path.lock().await = Some(path);
-                                    *state_setup.current_branch_id.lock().await = Some(branch.id);
-                                    *state_setup.current_branch_name.lock().await =
-                                        Some("main".to_string());
-                                    tracing::info!("Created new save file");
-                                }
-                            }
-                            Err(e) => {
-                                tracing::warn!("Failed to create save file: {}", e);
-                            }
-                        }
-                    } else {
-                        // All saves are locked by other instances.
-                        // Show the save picker so the user can choose or create a new ledger.
-                        tracing::info!(
-                            "All {} save file(s) are locked by other instances — opening save picker",
-                            saves.len()
-                        );
-                        let _ = handle.emit(events::EVENT_SAVE_PICKER, ());
-                    }
-                }
-
-                // ── Background ticks ─────────────────────────────────────────
-
-                // Event bus fan-in: subscribe to world.event_bus and buffer the
-                // last N events in AppState.game_events for the debug panel.
-                {
-                    let state_events = Arc::clone(&state_setup);
-                    let token_events = state_setup.shutdown_token.clone();
-                    let mut rx = {
-                        let world = state_events.world.lock().await;
-                        world.event_bus.subscribe()
-                    };
-                    tokio::spawn(async move {
-                        loop {
-                            tokio::select! {
-                                _ = token_events.cancelled() => break,
-                                result = rx.recv() => {
-                                    match result {
-                                        Ok(evt) => {
-                                            let mut buf = state_events.game_events.lock().await;
-                                            if buf.len() >= DEBUG_EVENT_CAPACITY {
-                                                buf.pop_front();
-                                            }
-                                            buf.push_back(evt);
-                                        }
-                                        Err(tokio::sync::broadcast::error::RecvError::Lagged(_)) => {
-                                            continue;
-                                        }
-                                        Err(tokio::sync::broadcast::error::RecvError::Closed) => {
-                                            break;
-                                        }
-                                    }
-                                }
-                            }
-                        }
-                    });
-                }
-
-                // Idle tick: emit world snapshot and run world/NPC ticks every 5 seconds.
-                // The GameClock already flows via speed_factor — no manual advance needed.
-                //
-                // Lock ordering: `world` → `npc_manager` → `debug_events`. Both
-                // `world` and `npc_manager` are acquired once at the top of each
-                // iteration and held through the entire body to avoid any window
-                // where a command handler could sneak in between them and race
-                // the tick (see the AppState lock ordering contract).
-                let state_tick = Arc::clone(&state_setup);
-                let handle_tick = handle.clone();
-                let token_tick = state_setup.shutdown_token.clone();
-                tokio::spawn(async move {
-                    let mut last_palette: Option<parish_palette::RawPalette> = None;
-                    loop {
-                        tokio::select! {
-                            _ = token_tick.cancelled() => break,
-                            _ = tokio::time::sleep(Duration::from_secs(5)) => {}
-                        }
-
-                        let mut world = state_tick.world.lock().await;
-                        let mut npc_mgr = state_tick.npc_manager.lock().await;
-
-                        // Emit a fresh world snapshot to the frontend.
-                        {
-                            let transport = state_tick.transport.default_mode();
-                            let snapshot = crate::commands::get_world_snapshot_inner(
-                                &world,
-                                transport,
-                                Some(&npc_mgr),
-                                &state_tick.pronunciations,
-                            );
-                            let _ = handle_tick.emit(events::EVENT_WORLD_UPDATE, snapshot);
-                            // Emit current time-of-day palette
-                            {
-                                use chrono::Timelike;
-                                use parish_palette::compute_palette;
-                                let now = world.clock.now();
-                                let raw = compute_palette(now.hour(), now.minute());
-                                if last_palette != Some(raw) {
-                                    let _ = handle_tick.emit(
-                                        events::EVENT_THEME_UPDATE,
-                                        ThemePalette::from(raw),
-                                    );
-                                    last_palette = Some(raw);
-                                }
-                            }
-                        }
-                        {
-                            // Tick weather engine
-                            let season = world.clock.season();
-                            let now = world.clock.now();
-                            // Scope thread_rng tightly so it is dropped before any await.
-                            let new_weather_opt = {
-                                let mut rng = rand::rng();
-                                world.weather_engine.tick(now, season, &mut rng)
-                            };
-                            {
-                                if let Some(new_weather) = new_weather_opt {
-                                    let old = world.weather;
-                                    world.weather = new_weather;
-                                    world.event_bus.publish(
-                                        parish_core::world::events::GameEvent::WeatherChanged {
-                                            new_weather: new_weather.to_string(),
-                                            timestamp: world.clock.now(),
-                                        },
-                                    );
-                                    tracing::info!(old = %old, new = %new_weather, "Weather changed");
-                                    // Emit weather debug event
-                                    let mut debug_events =
-                                        state_tick.debug_events.lock().await;
-                                    if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                        debug_events.pop_front();
-                                    }
-                                    debug_events.push_back(DebugEvent {
-                                        timestamp: String::new(),
-                                        category: "weather".to_string(),
-                                        message: format!(
-                                            "Weather: {} → {}",
-                                            old, new_weather
-                                        ),
-                                    });
-                                }
-                            }
-
-                            let schedule_events =
-                                npc_mgr.tick_schedules(&world.clock, &world.graph, world.weather);
-                            let tier_transitions = npc_mgr.assign_tiers(&world, &[]);
-
-                            // Banshee tick — herald and finalise doomed NPCs.
-                            // Default-on; kill-switched by the `banshee` feature flag.
-                            let banshee_enabled = {
-                                let cfg = state_tick.config.lock().await;
-                                !cfg.flags.is_disabled("banshee")
-                            };
-                            let banshee_report = if banshee_enabled {
-                                let world_ref = &mut *world;
-                                npc_mgr.tick_banshee(
-                                    &world_ref.clock,
-                                    &world_ref.graph,
-                                    &mut world_ref.text_log,
-                                    &world_ref.event_bus,
-                                    world_ref.player_location,
-                                )
-                            } else {
-                                parish_core::npc::banshee::BansheeReport::default()
-                            };
-                            if !banshee_report.is_empty() {
-                                let mut debug_events =
-                                    state_tick.debug_events.lock().await;
-                                if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                    debug_events.pop_front();
-                                }
-                                debug_events.push_back(DebugEvent {
-                                    timestamp: world.clock.now().format("%H:%M %Y-%m-%d").to_string(),
-                                    category: "banshee".to_string(),
-                                    message: format!(
-                                        "{} wail(s), {} death(s)",
-                                        banshee_report.wails.len(),
-                                        banshee_report.deaths.len()
-                                    ),
-                                });
-                            }
-
-                            // Log schedule events and tier transitions to debug panel
-                            if !schedule_events.is_empty() || !tier_transitions.is_empty() {
-                                let ts =
-                                    world.clock.now().format("%H:%M %Y-%m-%d").to_string();
-                                let mut debug_events = state_tick.debug_events.lock().await;
-                                for evt in &schedule_events {
-                                    if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                        debug_events.pop_front();
-                                    }
-                                    debug_events.push_back(DebugEvent {
-                                        timestamp: ts.clone(),
-                                        category: "schedule".to_string(),
-                                        message: evt.debug_string(),
-                                    });
-                                }
-                                for tt in &tier_transitions {
-                                    if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                        debug_events.pop_front();
-                                    }
-                                    let direction =
-                                        if tt.promoted { "promoted" } else { "demoted" };
-                                    debug_events.push_back(DebugEvent {
-                                        timestamp: ts.clone(),
-                                        category: "tier".to_string(),
-                                        message: format!(
-                                            "{} {} {:?} → {:?}",
-                                            tt.npc_name, direction, tt.old_tier, tt.new_tier,
-                                        ),
-                                    });
-                                }
-                            }
-
-                            // Propagate gossip between co-located Tier 2 NPCs
-                            // Scope thread_rng tightly so it is dropped before any await.
-                            let total_gossip = if !world.gossip_network.is_empty() {
-                                let groups = npc_mgr.tier2_groups();
-                                let mut rng = rand::rng();
-                                let mut total = 0usize;
-                                for npc_ids in groups.values() {
-                                    if npc_ids.len() >= 2 {
-                                        total +=
-                                            parish_core::npc::ticks::propagate_gossip_at_location(
-                                                npc_ids,
-                                                &mut world.gossip_network,
-                                                &mut rng,
-                                            );
-                                    }
-                                }
-                                total
-                            } else {
-                                0
-                            };
-                            {
-                                if total_gossip > 0 {
-                                    let mut debug_events =
-                                        state_tick.debug_events.lock().await;
-                                    if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                        debug_events.pop_front();
-                                    }
-                                    debug_events.push_back(DebugEvent {
-                                        timestamp: String::new(),
-                                        category: "gossip".to_string(),
-                                        message: format!(
-                                            "{} rumor(s) spread among co-located NPCs",
-                                            total_gossip
-                                        ),
-                                    });
-                                }
-                            }
-
-                            // Dispatch Tier 4 rules engine if enough game time has elapsed.
-                            // tick_tier4 is sub-ms CPU work; runs inline inside the lock scope.
-                            if npc_mgr.needs_tier4_tick(now) {
-                                let tier4_ids: std::collections::HashSet<parish_core::npc::NpcId> =
-                                    npc_mgr.tier4_npcs().into_iter().collect();
-                                let events = {
-                                    let mut tier4_refs: Vec<&mut parish_core::npc::Npc> = npc_mgr
-                                        .npcs_mut()
-                                        .values_mut()
-                                        .filter(|n| tier4_ids.contains(&n.id))
-                                        .collect();
-                                    let game_date = now.date_naive();
-                                    let mut rng = rand::rng();
-                                    parish_core::npc::tier4::tick_tier4(
-                                        &mut tier4_refs,
-                                        season,
-                                        game_date,
-                                        &mut rng,
-                                    )
-                                };
-                                let game_events = npc_mgr.apply_tier4_events(&events, now, banshee_enabled);
-                                // Collect per-event descriptions before publishing.
-                                let life_descriptions: Vec<String> = game_events
-                                    .iter()
-                                    .filter_map(|ge| {
-                                        if let parish_core::world::events::GameEvent::LifeEvent {
-                                            description,
-                                            ..
-                                        } = ge
-                                        {
-                                            Some(description.clone())
-                                        } else {
-                                            None
-                                        }
-                                    })
-                                    .collect();
-                                for evt in game_events {
-                                    world.event_bus.publish(evt);
-                                }
-                                npc_mgr.record_tier4_tick(now);
-                                let mut debug_events = state_tick.debug_events.lock().await;
-                                // Per-event life_event entries
-                                for desc in &life_descriptions {
-                                    if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                        debug_events.pop_front();
-                                    }
-                                    debug_events.push_back(DebugEvent {
-                                        timestamp: String::new(),
-                                        category: "life_event".to_string(),
-                                        message: desc.clone(),
-                                    });
-                                }
-                                // Aggregate tier4 entry
-                                if debug_events.len() >= crate::DEBUG_EVENT_CAPACITY {
-                                    debug_events.pop_front();
-                                }
-                                debug_events.push_back(DebugEvent {
-                                    timestamp: String::new(),
-                                    category: "tier4".to_string(),
-                                    message: format!("Tier 4 tick: {} events", events.len()),
-                                });
-                            }
-
-                            // Dispatch Tier 3 batch LLM simulation for distant NPCs.
-                            // The LLM call can take 10-30 s, so we spawn a detached task
-                            // and release the world/npc_mgr locks before awaiting.
-                            if npc_mgr.needs_tier3_tick(now)
-                                && !npc_mgr.tier3_in_flight()
-                            {
-                                use parish_core::npc::ticks::tier3_snapshot_from_npc;
-                                use parish_core::npc::ticks::Tier3Snapshot;
-
-                                let tier3_ids = npc_mgr.tier3_npcs();
-                                let snapshots: Vec<Tier3Snapshot> = tier3_ids
-                                    .iter()
-                                    .filter_map(|id| npc_mgr.get(*id))
-                                    .map(|npc| tier3_snapshot_from_npc(npc, &world.graph))
-                                    .collect();
-
-                                if !snapshots.is_empty() {
-                                    let time_desc =
-                                        world.clock.time_of_day().to_string();
-                                    let weather_str = world.weather.to_string();
-                                    let season_str =
-                                        format!("{:?}", world.clock.season());
-                                    let hours = 24u32;
-
-                                    npc_mgr.set_tier3_in_flight(true);
-
-                                    let state_t3 = Arc::clone(&state_tick);
-                                    tokio::spawn(async move {
-                                        // Briefly lock to clone the queue + resolve the model.
-                                        // NOTE: queue submissions go through the base worker
-                                        // client; per-category Simulation overrides are not
-                                        // honored for batch inference. TODO: per-category
-                                        // routing through the queue worker.
-                                        let (queue_opt, model) = {
-                                            let cfg = state_t3.config.lock().await;
-                                            let queue_guard =
-                                                state_t3.inference_queue.lock().await;
-                                            let queue = queue_guard.clone();
-                                            let model = cfg
-                                                .category_model
-                                                .get(&parish_core::config::InferenceCategory::Simulation)
-                                                .cloned()
-                                                .unwrap_or_else(|| cfg.model_name.clone());
-                                            (queue, model)
-                                        };
-
-                                        let Some(queue) = queue_opt else {
-                                            state_t3
-                                                .npc_manager
-                                                .lock()
-                                                .await
-                                                .set_tier3_in_flight(false);
-                                            return;
-                                        };
-
-                                        let ctx = parish_core::npc::ticks::Tier3Context {
-                                            snapshots: &snapshots,
-                                            queue: &queue,
-                                            model: &model,
-                                            time_desc: &time_desc,
-                                            weather: &weather_str,
-                                            season: &season_str,
-                                            hours,
-                                            batch_size: 0,
-                                            language: &state_t3.language_settings,
-                                        };
-
-                                        let result =
-                                            parish_core::npc::ticks::tick_tier3(&ctx)
-                                                .await;
-
-                                        // Re-acquire locks to apply updates.
-                                        // Lock ordering: `world` → `npc_manager`
-                                        // (matches the documented contract and the
-                                        // main tick at lib.rs:955-956).  Acquiring
-                                        // npc_manager first while a concurrent main
-                                        // tick holds world would deadlock (#337).
-                                        let world = state_t3.world.lock().await;
-                                        let mut npc_mgr =
-                                            state_t3.npc_manager.lock().await;
-                                        let game_time = world.clock.now();
-
-                                        match result {
-                                            Ok(updates) => {
-                                                let _events =
-                                                    parish_core::npc::ticks::apply_tier3_updates(
-                                                        &updates,
-                                                        npc_mgr.npcs_mut(),
-                                                        &world.graph,
-                                                        game_time,
-                                                    );
-                                                npc_mgr.record_tier3_tick(game_time);
-                                                tracing::debug!(
-                                                    "Tier 3 tick: {} updates applied",
-                                                    updates.len()
-                                                );
-
-                                                let mut debug_events =
-                                                    state_t3.debug_events.lock().await;
-                                                if debug_events.len()
-                                                    >= crate::DEBUG_EVENT_CAPACITY
-                                                {
-                                                    debug_events.pop_front();
-                                                }
-                                                debug_events.push_back(DebugEvent {
-                                                    timestamp: String::new(),
-                                                    category: "tier3".to_string(),
-                                                    message: format!(
-                                                        "Tier 3 tick: {} updates",
-                                                        updates.len()
-                                                    ),
-                                                });
-                                            }
-                                            Err(e) => {
-                                                tracing::warn!(
-                                                    "Tier 3 tick failed: {}",
-                                                    e
-                                                );
-                                            }
-                                        }
-
-                                        npc_mgr.set_tier3_in_flight(false);
-                                    });
-                                }
-                            }
-
-                            // Dispatch Tier 2 background simulation for nearby NPCs.
-                            // Submits one LLM call per location group via the priority queue
-                            // (Background lane, yields to Tier 1 dialogue).
-                            if npc_mgr.needs_tier2_tick(now)
-                                && !npc_mgr.tier2_in_flight()
-                            {
-                                use parish_core::npc::ticks::{
-                                    Tier2Group, npc_snapshot_from_npc,
-                                };
-
-                                let groups_map = npc_mgr.tier2_groups();
-                                if !groups_map.is_empty() {
-                                    // Build owned snapshots inside the lock scope.
-                                    let groups: Vec<Tier2Group> = groups_map
-                                        .into_iter()
-                                        .filter_map(|(loc, npc_ids)| {
-                                            let location_name = world
-                                                .graph
-                                                .get(loc)
-                                                .map(|d| d.name.clone())
-                                                .unwrap_or_else(|| {
-                                                    format!("Location {}", loc.0)
-                                                });
-                                            let npcs: Vec<_> = npc_ids
-                                                .iter()
-                                                .filter_map(|id| npc_mgr.get(*id))
-                                                .map(npc_snapshot_from_npc)
-                                                .collect();
-                                            if npcs.is_empty() {
-                                                return None;
-                                            }
-                                            Some(Tier2Group {
-                                                location: loc,
-                                                location_name,
-                                                npcs,
-                                            })
-                                        })
-                                        .collect();
-
-                                    if !groups.is_empty() {
-                                        let time_desc =
-                                            world.clock.time_of_day().to_string();
-                                        let weather_str = world.weather.to_string();
-
-                                        npc_mgr.set_tier2_in_flight(true);
-
-                                        let state_t2 = Arc::clone(&state_tick);
-                                        tokio::spawn(async move {
-                                            // Briefly lock to clone the queue + resolve model.
-                                            // NOTE: queue submissions go through the base worker
-                                            // client; per-category Simulation overrides are not
-                                            // honored for batch inference. TODO: per-category
-                                            // routing through the queue worker.
-                                            let (queue_opt, model) = {
-                                                let cfg = state_t2.config.lock().await;
-                                                let queue_guard =
-                                                    state_t2.inference_queue.lock().await;
-                                                let queue = queue_guard.clone();
-                                                let model = cfg
-                                                    .category_model
-                                                    .get(&parish_core::config::InferenceCategory::Simulation)
-                                                    .cloned()
-                                                    .unwrap_or_else(|| {
-                                                        cfg.model_name.clone()
-                                                    });
-                                                (queue, model)
-                                            };
-
-                                            let Some(queue) = queue_opt else {
-                                                state_t2
-                                                    .npc_manager
-                                                    .lock()
-                                                    .await
-                                                    .set_tier2_in_flight(false);
-                                                return;
-                                            };
-
-                                            // Submit each group sequentially (one LLM call
-                                            // per group, single connection).
-                                            let mut events = Vec::new();
-                                            for group in &groups {
-                                                if let Some(evt) =
-                                                    parish_core::npc::ticks::run_tier2_for_group(
-                                                        &queue,
-                                                        &model,
-                                                        group,
-                                                        &time_desc,
-                                                        &weather_str,
-                                                        &state_t2.language_settings,
-                                                    )
-                                                    .await
-                                                {
-                                                    events.push(evt);
-                                                }
-                                            }
-
-                                            // Re-acquire locks to apply events.
-                                            // Lock ordering: `world` → `npc_manager`
-                                            // (matches the documented contract and the
-                                            // main tick at lib.rs:955-956).  Acquiring
-                                            // npc_manager first while a concurrent main
-                                            // tick holds world would deadlock (#337).
-                                            let mut world = state_t2.world.lock().await;
-                                            let mut npc_mgr =
-                                                state_t2.npc_manager.lock().await;
-                                            let game_time = world.clock.now();
-
-                                            for event in &events {
-                                                let _dbg =
-                                                    parish_core::npc::ticks::apply_tier2_event(
-                                                        event,
-                                                        npc_mgr.npcs_mut(),
-                                                        game_time,
-                                                    );
-                                                // Push gossip so it can propagate to other NPCs.
-                                                parish_core::npc::ticks::create_gossip_from_tier2_event(
-                                                    event,
-                                                    &mut world.gossip_network,
-                                                    game_time,
-                                                );
-                                            }
-                                            npc_mgr.record_tier2_tick(game_time);
-                                            npc_mgr.set_tier2_in_flight(false);
-
-                                            let mut debug_events =
-                                                state_t2.debug_events.lock().await;
-                                            if debug_events.len()
-                                                >= crate::DEBUG_EVENT_CAPACITY
-                                            {
-                                                debug_events.pop_front();
-                                            }
-                                            debug_events.push_back(DebugEvent {
-                                                timestamp: String::new(),
-                                                category: "tier2".to_string(),
-                                                message: format!(
-                                                    "Tier 2 tick: {} events from {} groups",
-                                                    events.len(),
-                                                    groups.len()
-                                                ),
-                                            });
-                                        });
-                                    }
-                                }
-                            }
-                        }
-
-                        // Advance the generation counter so handle_game_input can
-                        // detect TOCTOU races (see issue #283).
-                        world.increment_tick_generation();
-                    }
-                });
-
-                // Inactivity tick: drive idle banter and auto-pause.
-                let state_idle = Arc::clone(&state_setup);
-                let handle_idle = handle.clone();
-                let token_idle = state_setup.shutdown_token.clone();
-                tokio::spawn(async move {
-                    loop {
-                        tokio::select! {
-                            _ = token_idle.cancelled() => break,
-                            _ = tokio::time::sleep(Duration::from_secs(1)) => {}
-                        }
-                        crate::commands::tick_inactivity(&state_idle, &handle_idle).await;
-                    }
-                });
-                // Debug tick: emit debug snapshot every 2 seconds.
-                //
-                // Snapshot each piece of state with a brief, non-overlapping
-                // lock window to avoid holding all 5+ locks simultaneously
-                // (#105, #282). Lock order: world → npc_manager →
-                // inference_queue → config → debug_events → game_events →
-                // inference_log (#483).
-                let state_debug = Arc::clone(&state_setup);
-                let handle_debug = handle.clone();
-                let token_debug = state_setup.shutdown_token.clone();
-                tokio::spawn(async move {
-                    loop {
-                        tokio::select! {
-                            _ = token_debug.cancelled() => break,
-                            _ = tokio::time::sleep(Duration::from_secs(2)) => {}
-                        }
-
-                        // 1. Peek inference_queue presence first (#483).
-                        let has_inference_queue =
-                            state_debug.inference_queue.lock().await.is_some();
-
-                        // 2. Clone config fields — drop the lock immediately.
-                        let (
-                            provider_name,
-                            model_name,
-                            base_url,
-                            cloud_provider,
-                            cloud_model,
-                            improv_enabled,
-                            categories,
-                        ) = {
-                            let config = state_debug.config.lock().await;
-                            (
-                                config.provider_name.clone(),
-                                config.model_name.clone(),
-                                config.base_url.clone(),
-                                config.cloud_provider_name.clone(),
-                                config.cloud_model_name.clone(),
-                                config.improv_enabled,
-                                parish_core::debug_snapshot::build_inference_categories(&config),
-                            )
-                        };
-
-                        // 3. Clone debug_events ring buffer — drop immediately.
-                        let debug_events_snapshot: std::collections::VecDeque<
-                            parish_core::debug_snapshot::DebugEvent,
-                        > = state_debug
-                            .debug_events
-                            .lock()
-                            .await
-                            .iter()
-                            .cloned()
-                            .collect();
-
-                        // 4. Clone game_events ring buffer — drop immediately.
-                        let game_events_snapshot: std::collections::VecDeque<
-                            parish_core::world::events::GameEvent,
-                        > = state_debug
-                            .game_events
-                            .lock()
-                            .await
-                            .iter()
-                            .cloned()
-                            .collect();
-
-                        // 5. Clone inference log — drop immediately.
-                        let call_log: Vec<parish_core::debug_snapshot::InferenceLogEntry> =
-                            state_debug
-                                .inference_log
-                                .lock()
-                                .await
-                                .iter()
-                                .cloned()
-                                .collect();
-
-                        // Build InferenceDebug from cloned data (no locks held).
-                        let inference = InferenceDebug {
-                            provider_name,
-                            model_name,
-                            base_url,
-                            cloud_provider,
-                            cloud_model,
-                            has_queue: has_inference_queue,
-                            reaction_req_id: parish_core::game_session::reaction_req_id_peek(),
-                            improv_enabled,
-                            call_log,
-                            categories,
-                            configured_providers: parish_core::debug_snapshot::build_configured_providers(),
-                        };
-
-                        // 6. Acquire world and npc_manager (canonical order)
-                        // only for the pure-read snapshot build, then release.
-                        let world = state_debug.world.lock().await;
-                        let npc_manager = state_debug.npc_manager.lock().await;
-                        let snapshot = parish_core::debug_snapshot::build_debug_snapshot(
-                            &world,
-                            &npc_manager,
-                            &debug_events_snapshot,
-                            &game_events_snapshot,
-                            &inference,
-                            &parish_core::debug_snapshot::AuthDebug::disabled(),
-                        );
-                        drop(npc_manager);
-                        drop(world);
-
-                        let _ = handle_debug.emit(events::EVENT_DEBUG_UPDATE, snapshot);
-                    }
-                });
-
-                // Autosave tick: save snapshot every AUTOSAVE_INTERVAL_SECS (if a save file is active)
-                let state_autosave = Arc::clone(&state_setup);
-                let token_autosave = state_setup.shutdown_token.clone();
-                tokio::spawn(async move {
-                    loop {
-                        tokio::select! {
-                            _ = token_autosave.cancelled() => break,
-                            _ = tokio::time::sleep(Duration::from_secs(AUTOSAVE_INTERVAL_SECS)) => {}
-                        }
-
-                        // Only autosave if a save file and branch are active
-                        let save_path = state_autosave.save_path.lock().await.clone();
-                        let branch_id = *state_autosave.current_branch_id.lock().await;
-
-                        if let (Some(path), Some(bid)) = (save_path, branch_id) {
-                            let world = state_autosave.world.lock().await;
-                            let npc_manager = state_autosave.npc_manager.lock().await;
-                            let snapshot =
-                                parish_core::persistence::snapshot::GameSnapshot::capture(
-                                    &world,
-                                    &npc_manager,
-                                );
-                            drop(npc_manager);
-                            drop(world);
-
-                            match parish_core::persistence::Database::open(&path) {
-                                Ok(db) => match db.save_snapshot(bid, &snapshot) {
-                                    Ok(_) => tracing::debug!("Autosave complete"),
-                                    Err(e) => tracing::warn!("Autosave failed: {}", e),
-                                },
-                                Err(e) => tracing::warn!("Autosave DB open failed: {}", e),
-                            }
-                        }
-                    }
-                });
+                setup::init_inference_queue(&state_setup).await;
+                setup::init_persistence(&handle, &state_setup).await;
+                setup::spawn_event_bus_fanin(&state_setup).await;
+                setup::spawn_world_tick(handle.clone(), Arc::clone(&state_setup));
+                setup::spawn_inactivity_tick(handle.clone(), Arc::clone(&state_setup));
+                setup::spawn_debug_tick(handle.clone(), Arc::clone(&state_setup));
+                setup::spawn_autosave_tick(Arc::clone(&state_setup));
             });
 
             Ok(())
